Move the search results toggle behavior to CompareWithBase
The goal for this commit is that the "toggle behavior" is the same for
both the SearchView and the ResultsView.

Indeed previously we had a different behavior depending on the view: in
the SearchView we used to dispatch an action to change the
"checkedRevisions" state, but in the ResultsView we'd pass the changed
array to the parent, up to CompareWithBase.

Now everything is handled at the CompareWithBase component level, with
no reference to the "checkedRevisions" state.

Therefore this commit contains 3 different related changes:
1. The "toggle search result" behavior moves to `CompareWithBase`.
   Previously this logic was in a hook, called from `SearchResultsList`.
2. All changes to the "checkedRevisions" state are removed from the
   toggle logic.
3. The callbacks take the toggled item instead of the changed array. The
   toggling logic is implemented in "CompareWithBase".