Use the hasNonEditableState prop to define the default value for the two "in progress" boolean states
The "inProgress" boolean states will now control whether the different
parts of the form are editable.

The default values for the two inProgress boolean states in
CompareWithBase is now derived from the prop hasNonEditableState.
Indeed, the inProgress state and revs will be used in a following commit
as the data displayed to the user. In the Search View, they're editable
right away, but in the Results View the user has to click Edit first.